requirements.txt
kurigram==2.2.12
requests==2.32.5
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
llama-cpp-python==0.3.16
httpx==0.28.1